restructure so `CameraCalibrator` is parent to `BoardDetection`
probably going to want to move the cam_managing stuff one step outwards and call it 'SessionManager' or something
